Number Guessing Game
1--NUMBER GUESSING GAME

The system generates a random number from a given range, say 1 to 100. The user is prompted to enter their given number in a displayed dialogue box. The computer then tells if the entered number matches the guesses number or it is higher/lower than the generated number. The game continues under the user guessing the number. It also includes: Limiting the number of attempts. Adding more rounds. Displaying score. Giving points based on the number of attempts.